+++ seatunnel-engine/seatunnel-engine-server/src/main/java/org/apache/seatunnel/engine/server/rest/service/SystemMonitoringService.java
@@ -1,0 +1,14 @@
+package org.apache.seatunnel.engine.server.rest.service;
+
+import com.hazelcast.internal.json.JsonArray;
+import com.hazelcast.spi.impl.NodeEngineImpl;
+
+public class SystemMonitoringService extends BaseService {
+    public SystemMonitoringService(NodeEngineImpl nodeEngine) {
+        super(nodeEngine);
+    }
+
+    public JsonArray getSystemMonitoringInformationJsonValues() {
+        return super.getSystemMonitoringInformationJsonValues();
+    }
+}
