+++ seatunnel-engine/seatunnel-engine-server/src/main/java/org/apache/seatunnel/engine/server/rest/servlet/JobInfoServlet.java
@@ -17,15 +17,8 @@
 
 package org.apache.seatunnel.engine.server.rest.servlet;
 
-import org.apache.seatunnel.api.common.metrics.JobMetrics;
-import org.apache.seatunnel.engine.common.Constant;
-import org.apache.seatunnel.engine.core.job.JobDAGInfo;
-import org.apache.seatunnel.engine.core.job.JobInfo;
-import org.apache.seatunnel.engine.server.master.JobHistoryService.JobState;
-import org.apache.seatunnel.engine.server.rest.RestConstant;
+import org.apache.seatunnel.engine.server.rest.service.JobInfoService;
 
-import com.hazelcast.internal.json.JsonObject;
-import com.hazelcast.map.IMap;
 import com.hazelcast.spi.impl.NodeEngineImpl;
 
 import javax.servlet.ServletException;
@@ -35,51 +28,27 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 public class JobInfoServlet extends BaseServlet {
+
+    private final JobInfoService jobInfoService;
+
     public JobInfoServlet(NodeEngineImpl nodeEngine) {
         super(nodeEngine);
+        this.jobInfoService = new JobInfoService(nodeEngine);
     }
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
 
-        String jobId = req.getPathInfo();
+        String jobIdStr = req.getPathInfo();
 
-        if (jobId != null && jobId.length() > 1) {
-            jobId = jobId.substring(1);
+        if (jobIdStr != null && jobIdStr.length() > 1) {
+            jobIdStr = jobIdStr.substring(1);
         } else {
             throw new IllegalArgumentException("The jobId must not be empty.");
         }
+        Long jobId = Long.valueOf(jobIdStr);
 
-        IMap<Object, Object> jobInfoMap =
-                nodeEngine.getHazelcastInstance().getMap(Constant.IMAP_RUNNING_JOB_INFO);
-        Object jobInfo = jobInfoMap.get(Long.valueOf(jobId));
-        IMap<Object, Object> finishedJobStateMap =
-                nodeEngine.getHazelcastInstance().getMap(Constant.IMAP_FINISHED_JOB_STATE);
-        Object finishedJobState = finishedJobStateMap.get(Long.valueOf(jobId));
-        if (jobInfo != null) {
-            writeJson(resp, convertToJson((JobInfo) jobInfo, Long.parseLong(jobId)));
-        } else if (finishedJobState != null) {
-            JobMetrics finishedJobMetrics =
-                    (JobMetrics)
-                            nodeEngine
-                                    .getHazelcastInstance()
-                                    .getMap(Constant.IMAP_FINISHED_JOB_METRICS)
-                                    .get(Long.valueOf(jobId));
-            JobDAGInfo finishedJobDAGInfo =
-                    (JobDAGInfo)
-                            nodeEngine
-                                    .getHazelcastInstance()
-                                    .getMap(Constant.IMAP_FINISHED_JOB_VERTEX_INFO)
-                                    .get(Long.valueOf(jobId));
-            writeJson(
-                    resp,
-                    getJobInfoJson(
-                            (JobState) finishedJobState,
-                            finishedJobMetrics.toJsonString(),
-                            finishedJobDAGInfo));
-        } else {
-            writeJson(resp, new JsonObject().add(RestConstant.JOB_ID, jobId));
-        }
+        writeJson(resp, jobInfoService.getJobInfoJson(jobId));
     }
 }
