+++ seatunnel-engine/seatunnel-engine-server/src/main/java/org/apache/seatunnel/engine/server/rest/servlet/SubmitJobServlet.java
@@ -17,16 +17,10 @@
 
 package org.apache.seatunnel.engine.server.rest.servlet;
 
-import org.apache.seatunnel.shade.com.typesafe.config.Config;
+import org.apache.seatunnel.engine.server.rest.service.JobInfoService;
 
-import org.apache.seatunnel.engine.server.SeaTunnelServer;
-import org.apache.seatunnel.engine.server.rest.RestConstant;
-import org.apache.seatunnel.engine.server.utils.RestUtil;
-
-import com.hazelcast.internal.json.JsonObject;
 import com.hazelcast.spi.impl.NodeEngineImpl;
 
-import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
@@ -34,26 +28,18 @@ import java.io.IOException;
 import java.util.Map;
 
 public class SubmitJobServlet extends BaseServlet {
+    private final JobInfoService jobInfoService;
+
     public SubmitJobServlet(NodeEngineImpl nodeEngine) {
         super(nodeEngine);
+        this.jobInfoService = new JobInfoService(nodeEngine);
     }
 
     @Override
-    public void doPost(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
+    public void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException {
 
         Map<String, String> requestParams = getParameterMap(req);
 
-        if (Boolean.parseBoolean(requestParams.get(RestConstant.IS_START_WITH_SAVE_POINT))
-                && requestParams.get(RestConstant.JOB_ID) == null) {
-            throw new IllegalArgumentException("Please provide jobId when start with save point.");
-        }
-
-        Config config = RestUtil.buildConfig(requestHandle(requestBody(req)), false);
-
-        SeaTunnelServer seaTunnelServer = getSeaTunnelServer(false);
-        JsonObject jsonObject =
-                submitJobInternal(config, requestParams, seaTunnelServer, nodeEngine.getNode());
-        writeJson(resp, jsonObject);
+        writeJson(resp, jobInfoService.submitJob(requestParams, requestBody(req)));
     }
 }
