+++ seatunnel-engine/seatunnel-engine-server/src/main/java/org/apache/seatunnel/engine/server/rest/servlet/SubmitJobsServlet.java
@@ -17,51 +17,27 @@
 
 package org.apache.seatunnel.engine.server.rest.servlet;
 
-import org.apache.seatunnel.shade.com.typesafe.config.Config;
+import org.apache.seatunnel.engine.server.rest.service.JobInfoService;
 
-import org.apache.seatunnel.engine.server.SeaTunnelServer;
-import org.apache.seatunnel.engine.server.utils.RestUtil;
-
-import com.hazelcast.internal.json.JsonArray;
 import com.hazelcast.spi.impl.NodeEngineImpl;
-import scala.Tuple2;
 
-import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 
 public class SubmitJobsServlet extends BaseServlet {
+
+    private final JobInfoService jobInfoService;
+
     public SubmitJobsServlet(NodeEngineImpl nodeEngine) {
         super(nodeEngine);
+        this.jobInfoService = new JobInfoService(nodeEngine);
     }
 
     @Override
-    public void doPost(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
-        List<Tuple2<Map<String, String>, Config>> configTuples =
-                RestUtil.buildConfigList(requestHandle(requestBody(req)), false);
+    public void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException {
 
-        JsonArray jsonArray =
-                configTuples.stream()
-                        .map(
-                                tuple -> {
-                                    String urlParams = mapToUrlParams(tuple._1);
-                                    Map<String, String> requestParams = new HashMap<>();
-                                    RestUtil.buildRequestParams(requestParams, urlParams);
-                                    SeaTunnelServer seaTunnelServer = getSeaTunnelServer(false);
-                                    return submitJobInternal(
-                                            tuple._2,
-                                            requestParams,
-                                            seaTunnelServer,
-                                            nodeEngine.getNode());
-                                })
-                        .collect(JsonArray::new, JsonArray::add, JsonArray::add);
-
-        writeJson(resp, jsonArray);
+        writeJson(resp, jobInfoService.submitJobs(requestBody(req)));
     }
 }
